requirements.txt
fastapi
uvicorn[standard]
numpy
pillow
scikit-learn
python-multipart
